auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
pyotp==2.9.0
python-dotenv==1.0.0
passlib==1.7.4
